scripts/docker-bot/requirements.txt
aiohttp==3.9.3
boto3==1.34.149
discord.py==2.3.2
python_logging_loki==0.3.1
Requests==2.32.3
